INIT - works!
A plugin to allow for importing of tags through a formatted .txt file. 
The file should be formatted as such: "id=?;track=?;artist=?;...", where each initial value represents a valid beets key tag, and the value following the = represents the new value for that key. These pairs are broken up by semicolons. The line can be as long as you want it to be, as long as the keys are valid beets keys.